Character Movement
Some camera adjustments were needed to fix one character it is fixed now however to change characters we will need to implement this fix to the others. Got the animations working for forward and backward movement. Got camera shaking when running as well. Brought in many animations to work with to.